requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic

groq

langchain-core
langchain-community
langchain-text-splitters
chromadb

sentence-transformers

pypdf
python-docx
unstructured
